core/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
redis>=5.0
gunicorn>=21.0
python-dotenv>=1.0
celery>=5.3
django-cors-headers>=4.3
drf-spectacular>=0.27
PyJWT>=2.8
aiohttp>=3.9
cryptography>=41.0
requests>=2.31
# ML analytics via plexe (optional – install separately if LLM keys are available)
# plexe>=1.3.0
pandas>=2.0
pyarrow>=14.0
# Testing
pytest>=7.0
pytest-asyncio>=0.23
pytest-django>=4.7
